[mod/#6] HaphapTypography를 Type.kt로 이동
