+++ app/src/main/java/com/haphap/app/core/designsystem/theme/Type.kt
@@ -1,5 +1,6 @@
 package com.haphap.app.core.designsystem.theme
 
+import androidx.compose.runtime.Immutable
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
@@ -50,3 +51,58 @@ val c_r_12: TextStyle = pretendard(FontWeight.Normal, 12)
 val c_r_11: TextStyle = pretendard(FontWeight.Normal, 11)
 val c_sb_10: TextStyle = pretendard(FontWeight.SemiBold, 10)
 val c_r_10: TextStyle = pretendard(FontWeight.Normal, 10)
+
+@Immutable
+data class HaphapTypography(
+    val t_b_30: TextStyle,
+    val t_b_26: TextStyle,
+    val st_b_24: TextStyle,
+    val st_sb_24: TextStyle,
+    val st_b_22: TextStyle,
+    val st_b_20: TextStyle,
+    val b_b_18: TextStyle,
+    val b_sb_18: TextStyle,
+    val b_m_18: TextStyle,
+    val b_r_18: TextStyle,
+    val b_sb_16: TextStyle,
+    val b_r_16: TextStyle,
+    val b_b_14: TextStyle,
+    val b_sb_14: TextStyle,
+    val b_m_14: TextStyle,
+    val b_r_14: TextStyle,
+    val b_sb_13: TextStyle,
+    val c_b_12: TextStyle,
+    val c_sb_12: TextStyle,
+    val c_m_12: TextStyle,
+    val c_r_12: TextStyle,
+    val c_r_11: TextStyle,
+    val c_sb_10: TextStyle,
+    val c_r_10: TextStyle,
+)
+
+val defaultHaphapTypography = HaphapTypography(
+    t_b_30 = t_b_30,
+    t_b_26 = t_b_26,
+    st_b_24 = st_b_24,
+    st_sb_24 = st_sb_24,
+    st_b_22 = st_b_22,
+    st_b_20 = st_b_20,
+    b_b_18 = b_b_18,
+    b_sb_18 = b_sb_18,
+    b_m_18 = b_m_18,
+    b_r_18 = b_r_18,
+    b_sb_16 = b_sb_16,
+    b_r_16 = b_r_16,
+    b_b_14 = b_b_14,
+    b_sb_14 = b_sb_14,
+    b_m_14 = b_m_14,
+    b_r_14 = b_r_14,
+    b_sb_13 = b_sb_13,
+    c_b_12 = c_b_12,
+    c_sb_12 = c_sb_12,
+    c_m_12 = c_m_12,
+    c_r_12 = c_r_12,
+    c_r_11 = c_r_11,
+    c_sb_10 = c_sb_10,
+    c_r_10 = c_r_10,
+)
